Hugh update of the view
-Data point markers only show when you touch it
-Added drop shadow as an alternative
-Fixed a few rendering issues
-Increased performance
-Smoother lines
-Better scaling of the chart
-Source better commented and looks a bit better.

+++ src/com/example/linechart/DataPoints.java
@@ -3,14 +3,14 @@ package com.example.linechart;
 public class DataPoints
 {
   private String _info;
-  private int _point;
+  private float _point;
 
   public String getInfo()
   {
     return _info;
   }
 
-  public int getPoint()
+  public float getPoint()
   {
     return _point;
   }
